requirements.txt
ffmpeg==4.4.2
iopath==0.1.10
opencv-python==4.13.0.92
Pillow==12.1.1
pydub==0.25.1
scikit-image==0.25.2
soundfile==0.13.1
timm==1.0.25
torch==2.10.0
torchvision==2.10.0
torchaudio==0.25.0
torchcodec==0.10.0
tqdm==4.67.3
wand==0.7.0
yacs==0.1.8